requirements.txt
exif==1.4.0
Flask==2.2.2
Pillow==9.3.0
